ref: move custom useForm hook into shared/hooks
